[Core] BaseModel backend registry refactor
