+++ pkg/controller/v1beta1/basemodel/controller_test.go
@@ -16,11 +16,10 @@ import (
 	ctrl "sigs.k8s.io/controller-runtime"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 	ctrlclientfake "sigs.k8s.io/controller-runtime/pkg/client/fake"
-	"sigs.k8s.io/controller-runtime/pkg/event"
-	"sigs.k8s.io/controller-runtime/pkg/reconcile"
 
 	"sigs.k8s.io/ome/pkg/apis/ome/v1beta1"
 	"sigs.k8s.io/ome/pkg/constants"
+	"sigs.k8s.io/ome/pkg/controller/v1beta1/basemodel/shared"
 	"sigs.k8s.io/ome/pkg/modelagent"
 )
 
@@ -749,131 +748,6 @@ func TestClusterBaseModelReconcile(t *testing.T) {
 	}
 }
 
-func TestMapConfigMapToModelRequests(t *testing.T) {
-	g := gomega.NewGomegaWithT(t)
-	logger := ctrl.Log.WithName("test")
-
-	tests := []struct {
-		name          string
-		configMap     *corev1.ConfigMap
-		keyPrefix     string
-		isNamespaced  bool
-		expectedCount int
-		expectedFirst *types.NamespacedName
-	}{
-		{
-			name: "BaseModel mapping",
-			configMap: &corev1.ConfigMap{
-				ObjectMeta: metav1.ObjectMeta{
-					Name:      "test-node",
-					Namespace: constants.OMENamespace,
-				},
-				Data: map[string]string{
-					"default.basemodel.my-model":     `{"status":"Ready"}`,
-					"test-ns.basemodel.other-model":  `{"status":"Failed"}`,
-					"clusterbasemodel.cluster-model": `{"status":"Ready"}`, // Should be ignored
-				},
-			},
-			keyPrefix:     "basemodel",
-			isNamespaced:  true,
-			expectedCount: 2,
-			expectedFirst: nil, // Don't check specific order since map iteration is random
-		},
-		{
-			name: "ClusterBaseModel mapping",
-			configMap: &corev1.ConfigMap{
-				ObjectMeta: metav1.ObjectMeta{
-					Name:      "cluster-node",
-					Namespace: constants.OMENamespace,
-				},
-				Data: map[string]string{
-					"clusterbasemodel.global-model":    `{"status":"Ready"}`,
-					"clusterbasemodel.multi.part.name": `{"status":"InTransit"}`,
-					"basemodel.default.local-model":    `{"status":"Ready"}`, // Should be ignored
-				},
-			},
-			keyPrefix:     "clusterbasemodel",
-			isNamespaced:  false,
-			expectedCount: 2,
-			expectedFirst: nil, // Don't check specific order since map iteration is random
-		},
-		{
-			name: "Invalid JSON data",
-			configMap: &corev1.ConfigMap{
-				ObjectMeta: metav1.ObjectMeta{
-					Name:      "invalid-node",
-					Namespace: constants.OMENamespace,
-				},
-				Data: map[string]string{
-					"default.basemodel.broken-model": `{invalid json}`,
-					"default.basemodel.valid-model":  `{"status":"Ready"}`,
-				},
-			},
-			keyPrefix:     "basemodel",
-			isNamespaced:  true,
-			expectedCount: 1, // Only valid entry should be processed
-			expectedFirst: &types.NamespacedName{
-				Namespace: "default",
-				Name:      "valid-model",
-			},
-		},
-		{
-			name:          "Non-ConfigMap object",
-			configMap:     nil, // Will pass a different object type
-			keyPrefix:     "basemodel",
-			isNamespaced:  true,
-			expectedCount: 0,
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			var requests []reconcile.Request
-
-			if tt.configMap != nil {
-				if tt.isNamespaced {
-					// Test BaseModel mapping
-					reconciler := &BaseModelReconciler{Log: logger}
-					requests = reconciler.mapConfigMapToBaseModels(tt.configMap)
-				} else {
-					// Test ClusterBaseModel mapping
-					reconciler := &ClusterBaseModelReconciler{Log: logger}
-					requests = reconciler.mapConfigMapToClusterBaseModels(tt.configMap)
-				}
-			} else {
-				// Pass a non-ConfigMap object
-				reconciler := &BaseModelReconciler{Log: logger}
-				requests = reconciler.mapConfigMapToBaseModels(&corev1.Pod{})
-			}
-
-			g.Expect(requests).To(gomega.HaveLen(tt.expectedCount))
-
-			if tt.expectedFirst != nil && len(requests) > 0 {
-				g.Expect(requests[0].NamespacedName).To(gomega.Equal(*tt.expectedFirst))
-			} else if tt.expectedCount > 0 {
-				// Instead of checking order, verify that all expected requests are present
-				// For BaseModel mapping case
-				if tt.name == "BaseModel mapping" {
-					foundDefault := false
-					foundTestNs := false
-
-					for _, req := range requests {
-						if req.NamespacedName.Namespace == "default" && req.NamespacedName.Name == "my-model" {
-							foundDefault = true
-						}
-						if req.NamespacedName.Namespace == "test-ns" && req.NamespacedName.Name == "other-model" {
-							foundTestNs = true
-						}
-					}
-
-					g.Expect(foundDefault).To(gomega.BeTrue(), "Should find default.my-model")
-					g.Expect(foundTestNs).To(gomega.BeTrue(), "Should find test-ns.other-model")
-				}
-			}
-		})
-	}
-}
-
 func TestUpdateSpecWithConfig(t *testing.T) {
 	g := gomega.NewGomegaWithT(t)
 
@@ -972,333 +846,12 @@ func TestUpdateSpecWithConfig(t *testing.T) {
 
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			updated := updateSpecWithConfig(tt.initialSpec, tt.config)
+			updated := shared.UpdateSpecWithConfig(tt.initialSpec, tt.config)
 			g.Expect(updated).To(gomega.Equal(tt.expectUpdate))
 
 			if tt.validateSpec != nil && tt.initialSpec != nil {
 				tt.validateSpec(tt.initialSpec)
 			}
-		})
-	}
-}
-
-func TestCalculateLifecycleState(t *testing.T) {
-	g := gomega.NewGomegaWithT(t)
-
-	tests := []struct {
-		name          string
-		nodesReady    []string
-		nodesFailed   []string
-		expectedState v1beta1.LifeCycleState
-	}{
-		{
-			name:          "Ready state with ready nodes",
-			nodesReady:    []string{"node1", "node2"},
-			nodesFailed:   []string{},
-			expectedState: v1beta1.LifeCycleStateReady,
-		},
-		{
-			name:          "Ready state with mixed nodes",
-			nodesReady:    []string{"node1"},
-			nodesFailed:   []string{"node2"},
-			expectedState: v1beta1.LifeCycleStateReady,
-		},
-		{
-			name:          "Failed state with only failed nodes",
-			nodesReady:    []string{},
-			nodesFailed:   []string{"node1", "node2"},
-			expectedState: v1beta1.LifeCycleStateFailed,
-		},
-		{
-			name:          "InTransit state with no nodes",
-			nodesReady:    []string{},
-			nodesFailed:   []string{},
-			expectedState: v1beta1.LifeCycleStateInTransit,
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			state := calculateLifecycleState(tt.nodesReady, tt.nodesFailed)
-			g.Expect(state).To(gomega.Equal(tt.expectedState))
-		})
-	}
-}
-
-func TestCreateModelStatusConfigMapPredicate(t *testing.T) {
-	g := gomega.NewGomegaWithT(t)
-
-	predicate := createModelStatusConfigMapPredicate()
-
-	tests := []struct {
-		name     string
-		obj      client.Object
-		expected bool
-	}{
-		{
-			name: "Valid model status ConfigMap",
-			obj: &corev1.ConfigMap{
-				ObjectMeta: metav1.ObjectMeta{
-					Name:      "test-node",
-					Namespace: constants.OMENamespace,
-					Labels: map[string]string{
-						constants.ModelStatusConfigMapLabel: "true",
-					},
-				},
-			},
-			expected: true,
-		},
-		{
-			name: "ConfigMap in wrong namespace",
-			obj: &corev1.ConfigMap{
-				ObjectMeta: metav1.ObjectMeta{
-					Name:      "test-node",
-					Namespace: "wrong-namespace",
-					Labels: map[string]string{
-						constants.ModelStatusConfigMapLabel: "true",
-					},
-				},
-			},
-			expected: false,
-		},
-		{
-			name: "ConfigMap without label",
-			obj: &corev1.ConfigMap{
-				ObjectMeta: metav1.ObjectMeta{
-					Name:      "test-node",
-					Namespace: constants.OMENamespace,
-				},
-			},
-			expected: false,
-		},
-		{
-			name: "ConfigMap with wrong label value",
-			obj: &corev1.ConfigMap{
-				ObjectMeta: metav1.ObjectMeta{
-					Name:      "test-node",
-					Namespace: constants.OMENamespace,
-					Labels: map[string]string{
-						constants.ModelStatusConfigMapLabel: "false",
-					},
-				},
-			},
-			expected: false,
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			// Test CreateFunc
-			result := predicate.Create(event.TypedCreateEvent[client.Object]{Object: tt.obj})
-			g.Expect(result).To(gomega.Equal(tt.expected))
-
-			// Test UpdateFunc
-			result = predicate.Update(event.TypedUpdateEvent[client.Object]{ObjectNew: tt.obj})
-			g.Expect(result).To(gomega.Equal(tt.expected))
-
-			// Test DeleteFunc
-			result = predicate.Delete(event.TypedDeleteEvent[client.Object]{Object: tt.obj})
-			g.Expect(result).To(gomega.Equal(tt.expected))
-		})
-	}
-}
-
-func TestAddToSlice(t *testing.T) {
-	g := gomega.NewGomegaWithT(t)
-
-	tests := []struct {
-		name     string
-		slice    []string
-		item     string
-		expected []string
-	}{
-		{
-			name:     "Add to empty slice",
-			slice:    []string{},
-			item:     "item1",
-			expected: []string{"item1"},
-		},
-		{
-			name:     "Add new item",
-			slice:    []string{"item1", "item2"},
-			item:     "item3",
-			expected: []string{"item1", "item2", "item3"},
-		},
-		{
-			name:     "Don't add existing item",
-			slice:    []string{"item1", "item2"},
-			item:     "item1",
-			expected: []string{"item1", "item2"},
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			result := addToSlice(tt.slice, tt.item)
-			g.Expect(result).To(gomega.Equal(tt.expected))
-		})
-	}
-}
-
-func TestCreateNodeDeletionPredicate(t *testing.T) {
-	g := gomega.NewGomegaWithT(t)
-
-	pred := createNodeDeletionPredicate()
-
-	node := &corev1.Node{
-		ObjectMeta: metav1.ObjectMeta{
-			Name: "test-node",
-		},
-	}
-
-	// CreateFunc should return false
-	createResult := pred.Create(event.TypedCreateEvent[client.Object]{Object: node})
-	g.Expect(createResult).To(gomega.BeFalse(), "CreateFunc should return false")
-
-	// UpdateFunc should return false
-	updateResult := pred.Update(event.TypedUpdateEvent[client.Object]{ObjectNew: node, ObjectOld: node})
-	g.Expect(updateResult).To(gomega.BeFalse(), "UpdateFunc should return false")
-
-	// DeleteFunc should return true
-	deleteResult := pred.Delete(event.TypedDeleteEvent[client.Object]{Object: node})
-	g.Expect(deleteResult).To(gomega.BeTrue(), "DeleteFunc should return true")
-}
-
-func TestHandleNodeDeletion(t *testing.T) {
-	g := gomega.NewGomegaWithT(t)
-
-	// Create scheme
-	scheme := runtime.NewScheme()
-	g.Expect(v1beta1.AddToScheme(scheme)).NotTo(gomega.HaveOccurred())
-	g.Expect(corev1.AddToScheme(scheme)).NotTo(gomega.HaveOccurred())
-
-	tests := []struct {
-		name       string
-		nodeName   string
-		setupMocks func(client.Client)
-		validate   func(*testing.T, client.Client, string)
-	}{
-		{
-			name:     "Node deletion cleans up associated ConfigMap",
-			nodeName: "node-with-configmap",
-			setupMocks: func(c client.Client) {
-				// Create ome namespace
-				omeNamespace := &corev1.Namespace{
-					ObjectMeta: metav1.ObjectMeta{
-						Name: constants.OMENamespace,
-					},
-				}
-				err := c.Create(context.TODO(), omeNamespace)
-				g.Expect(err).NotTo(gomega.HaveOccurred())
-
-				// Create a model status ConfigMap for this node
-				configMap := &corev1.ConfigMap{
-					ObjectMeta: metav1.ObjectMeta{
-						Name:      "node-with-configmap",
-						Namespace: constants.OMENamespace,
-						Labels: map[string]string{
-							constants.ModelStatusConfigMapLabel: "true",
-						},
-					},
-					Data: map[string]string{
-						"clusterbasemodel.test-model": `{"status":"Ready"}`,
-					},
-				}
-				err = c.Create(context.TODO(), configMap)
-				g.Expect(err).NotTo(gomega.HaveOccurred())
-			},
-			validate: func(t *testing.T, c client.Client, nodeName string) {
-				// ConfigMap should be deleted
-				configMap := &corev1.ConfigMap{}
-				err := c.Get(context.TODO(), types.NamespacedName{
-					Namespace: constants.OMENamespace,
-					Name:      nodeName,
-				}, configMap)
-				g.Expect(errors.IsNotFound(err)).To(gomega.BeTrue(), "ConfigMap should be deleted")
-			},
-		},
-		{
-			name:     "Node deletion with no ConfigMap does nothing",
-			nodeName: "node-without-configmap",
-			setupMocks: func(c client.Client) {
-				// Create ome namespace but no ConfigMap
-				omeNamespace := &corev1.Namespace{
-					ObjectMeta: metav1.ObjectMeta{
-						Name: constants.OMENamespace,
-					},
-				}
-				err := c.Create(context.TODO(), omeNamespace)
-				g.Expect(err).NotTo(gomega.HaveOccurred())
-			},
-			validate: func(t *testing.T, c client.Client, nodeName string) {
-				// No ConfigMap to check - just ensure no error occurred
-				// The function should silently skip
-			},
-		},
-		{
-			name:     "Node deletion skips non-model-status ConfigMap",
-			nodeName: "node-with-other-configmap",
-			setupMocks: func(c client.Client) {
-				// Create ome namespace
-				omeNamespace := &corev1.Namespace{
-					ObjectMeta: metav1.ObjectMeta{
-						Name: constants.OMENamespace,
-					},
-				}
-				err := c.Create(context.TODO(), omeNamespace)
-				g.Expect(err).NotTo(gomega.HaveOccurred())
-
-				// Create a ConfigMap without model status label
-				configMap := &corev1.ConfigMap{
-					ObjectMeta: metav1.ObjectMeta{
-						Name:      "node-with-other-configmap",
-						Namespace: constants.OMENamespace,
-						// No model status label
-					},
-					Data: map[string]string{
-						"some-key": "some-value",
-					},
-				}
-				err = c.Create(context.TODO(), configMap)
-				g.Expect(err).NotTo(gomega.HaveOccurred())
-			},
-			validate: func(t *testing.T, c client.Client, nodeName string) {
-				// ConfigMap should NOT be deleted (it's not a model status ConfigMap)
-				configMap := &corev1.ConfigMap{}
-				err := c.Get(context.TODO(), types.NamespacedName{
-					Namespace: constants.OMENamespace,
-					Name:      nodeName,
-				}, configMap)
-				g.Expect(err).NotTo(gomega.HaveOccurred(), "ConfigMap should still exist")
-			},
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			c := ctrlclientfake.NewClientBuilder().
-				WithScheme(scheme).
-				Build()
-
-			tt.setupMocks(c)
-
-			log := ctrl.Log.WithName("test")
-
-			// Create the node object that was "deleted"
-			deletedNode := &corev1.Node{
-				ObjectMeta: metav1.ObjectMeta{
-					Name: tt.nodeName,
-				},
-			}
-
-			// Call handleNodeDeletion (shared function)
-			requests := handleNodeDeletion(context.TODO(), c, log, deletedNode)
-
-			// Should return nil (no reconcile requests needed)
-			g.Expect(requests).To(gomega.BeNil())
-
-			// Validate the result
-			tt.validate(t, c, tt.nodeName)
 		})
 	}
 }
